[MOVE]: Response Class Class폴더 이동
